fuse map view models
